refactor: remove shared icon because of generate ui's lack of shadow dom
