+++ apps/generate-ui-v2/src/components/show-more-divider.ts
@@ -25,13 +25,13 @@ export class ShowMoreDivider extends EditorContext(LitElement) {
           }"
         />
 
-        <div tabindex="0" aria-role="button" class="focus:ring-1 focus:ring-focusBorder focus:outline-none" @keydown="${
+        <div tabindex="0" aria-role="button" class="flex flex-row gap-2 leading-none focus:ring-1 focus:ring-focusBorder focus:outline-none" @keydown="${
           this.handleKeyEvent
         }">${
       this.showMore ? 'Show less options' : 'Show all options'
     } <icon-element icon="${
       this.showMore ? 'chevron-up' : 'chevron-down'
-    }" class="align-sub"></div>
+    }" class="self-center"></div>
       </div>
     `;
   }
