--- libs/shared/lit-utils/README.md
@@ -1,11 +1,0 @@
-# shared-lit-utils
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build shared-lit-utils` to build the library.
-
-## Running unit tests
-
-Run `nx test shared-lit-utils` to execute the unit tests via [Jest](https://jestjs.io).
