--- libs/shared/lit-utils/jest.config.ts
@@ -1,10 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'shared-lit-utils',
-  preset: '../../../jest.preset.js',
-  transform: {
-    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../coverage/libs/shared/lit-utils',
-};
