--- libs/shared/lit-utils/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@nx-console/shared/lit-utils",
-  "version": "0.0.1",
-  "type": "commonjs"
-}
