--- libs/shared/lit-utils/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/codicon-element';
