--- libs/shared/lit-utils/tsconfig.lib.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "outDir": "../../../dist/out-tsc",
-    "declaration": true,
-    "types": ["node"],
-    "useDefineForClassFields": false
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
-}
