+++ libs/vscode/nx-cloud-view/src/webview/components/ui/codicon-element.ts
